feat: factored out tests from registry
