+++ client/src/actors/registry.rs
@@ -156,74 +156,3 @@ impl Handler<messages::GetAllClients> for Registry {
         result
     }
 }
-
-#[cfg(test)]
-mod tests {
-
-    use super::*;
-
-    #[actix::test]
-    async fn test_insert_client() {
-        let registry = Registry::default().start();
-
-        for d in 'a'..'z' {
-            let format_str = format!("{}", d).repeat(24);
-            let id_str = format_str.as_str().as_bytes();
-            let n = registry
-                .send(messages::InsertClient {
-                    id: ClientId::load(id_str).unwrap(),
-                })
-                .await;
-
-            assert!(n.is_ok());
-        }
-    }
-
-    #[actix::test]
-    async fn test_get_client() {
-        let registry = Registry::default().start();
-
-        for d in 'a'..'z' {
-            let format_str = format!("{}", d).repeat(24);
-            let id_str = format_str.as_str().as_bytes();
-            assert!(registry
-                .send(messages::InsertClient {
-                    id: ClientId::load(id_str).unwrap(),
-                })
-                .await
-                .is_ok());
-        }
-
-        assert!(registry
-            .send(messages::GetClient {
-                id: ClientId::load("b".repeat(24).as_bytes()).unwrap(),
-            })
-            .await
-            .is_ok());
-    }
-
-    #[actix::test]
-    async fn test_remove_client() {
-        let registry = Registry::default().start();
-
-        for d in 'a'..'z' {
-            let format_str = format!("{}", d).repeat(24);
-            let id_str = format_str.as_str().as_bytes();
-            assert!(registry
-                .send(messages::InsertClient {
-                    id: ClientId::load(id_str).unwrap(),
-                })
-                .await
-                .is_ok());
-        }
-
-        if let Ok(result) = registry
-            .send(messages::RemoveClient {
-                id: ClientId::load("a".repeat(24).as_bytes()).unwrap(),
-            })
-            .await
-        {
-            assert!(result.is_ok())
-        }
-    }
-}
