+++ client/src/tests.rs
@@ -4,6 +4,7 @@
 #[allow(dead_code)]
 mod fresh;
 
+mod actor_tests;
 mod basic_tests;
 mod interface_tests;
 mod procedures_tests;
